+++ src/mediawiki.ts
@@ -107,7 +107,7 @@ export class FullMediaWiki extends MediaWiki {
 	override mediawiki(tags?: string[]): StreamParser<any> { // eslint-disable-line @typescript-eslint/no-explicit-any
 		const parser = super.mediawiki(tags);
 		parser.languageData = {
-			closeBrackets: {brackets: ['(', '[', '{', '"']} satisfies CloseBracketConfig,
+			closeBrackets: {brackets: ['(', '[', '{', '"'], before: ')]}>'} satisfies CloseBracketConfig,
 			autocomplete: this.completionSource,
 		};
 		return parser;